+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/scheduler/SchedulerPlugin.kt
@@ -5,8 +5,7 @@ import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.Plugin
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.PluginManager
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.base.BasePlugin
 import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.publishers.Publisher
-import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.scheduler.commands.EnableTimerCommand
-import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.scheduler.commands.GetSchedulesCommand
+import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.scheduler.commands.*
 import com.pengrad.telegrambot.TelegramBot
 import org.jetbrains.exposed.sql.SchemaUtils
 import org.jetbrains.exposed.sql.transactions.transaction
@@ -17,6 +16,7 @@ class SchedulerPlugin : Plugin {
 
     private lateinit var enableTimerCommand: EnableTimerCommand
     private lateinit var getSchedulesCommand: GetSchedulesCommand
+    private lateinit var disableTimerCommand: DisableTimerCommand
 
     private lateinit var scheduler: Scheduler
 
@@ -46,6 +46,10 @@ class SchedulerPlugin : Plugin {
             timerSchedulesTable,
             botWR,
             baseConfig.sourceChatId
+        )
+        disableTimerCommand = DisableTimerCommand(
+            timerSchedulesTable,
+            botWR
         )
     }
 }